% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{aggregate_CV_split_results}
\alias{aggregate_CV_split_results}
\title{The required aggregate_resample_run_results method is needed to fulfill the
metric object interface. This function is called by the cross-validator to 
aggregrate results across all cross-validation splits. Note: this function should not be
called by users of the package but instead is called interally by cross-validator
objects.}
\usage{
aggregate_CV_split_results(rm_obj, prediction_results)
}
\arguments{
\item{rm_obj}{The results metric object}

\item{prediction_results}{the prediction results to be aggregated over CV splits}
}
\description{
The required aggregate_resample_run_results method is needed to fulfill the
metric object interface. This function is called by the cross-validator to 
aggregrate results across all cross-validation splits. Note: this function should not be
called by users of the package but instead is called interally by cross-validator
objects.
}
