% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{aggregate_resample_run_results}
\alias{aggregate_resample_run_results}
\title{The required aggregate_resample_run_results method is needed to fulfill the
metric object interface. This function is called by the cross-validator to 
aggregrate results across all resample runs. Note: this function should not be
called by users of the package but instead is called interally by cross-validator
objects.}
\usage{
aggregate_resample_run_results(obj)
}
\arguments{
\item{obj}{The RM object}
}
\description{
The required aggregate_resample_run_results method is needed to fulfill the
metric object interface. This function is called by the cross-validator to 
aggregrate results across all resample runs. Note: this function should not be
called by users of the package but instead is called interally by cross-validator
objects.
}
