% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_main_results.R
\name{aggregate_resample_run_results.rm_main_results}
\alias{aggregate_resample_run_results.rm_main_results}
\title{A result metric (RM) method to aggregate results over resample runs}
\usage{
\method{aggregate_resample_run_results}{rm_main_results}(resample_run_results)
}
\arguments{
\item{resample_run_results}{The decoding results from all resample runs.}
}
\value{
A result metric object that contains the decoding results aggregated
across resample runs. This compressed final results can be plotted (often
by using the RM plot methods).
}
\description{
This is a function that must be implemented by all RM objects. This function
is called by the cross-validator to aggregate results across all resample
runs. This method should not be called directly but instead it is used
internally by the cross-validator (CV) object.
}
\seealso{
\code{\link[=rm_main_results]{rm_main_results()}}, \code{\link[=rm_confusion_matrix]{rm_confusion_matrix()}}
}
\keyword{internal}
