% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_DS.R
\name{basic_DS}
\alias{basic_DS}
\title{A basic datasource object}
\usage{
basic_DS(binned_file_name = NULL, var_to_decode = NULL,
  num_cv_splits = NULL, use_count_data = FALSE,
  num_label_repeats_per_cv_split = 1, label_levels_to_use = NULL,
  num_resample_sites = NULL, site_IDs_to_use = NULL,
  site_IDs_to_exclude = NULL,
  randomly_shuffled_labels_before_running = FALSE,
  create_simultaneously_recorded_populations = 0)
}
\description{
A datasource object takes data in binned format and returns training 
  and testing splits of the data that can be passed to a classifier. 
  This object uses \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{R6 package}
}
\section{basic_DS constructor}{


\describe{
\item{\code{basic_DS$new(binned_data, var_to_decode, num_cv_splits, use_count_data, num_label_repeats_per_cv_split)}}{
if successful, will return a new \code{basic_DS} object.
}}
}

