% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_num_level_repetitions.R
\name{calc_num_level_repetitions}
\alias{calc_num_level_repetitions}
\title{calc_num_level_repetitions}
\usage{
calc_num_level_repetitions(binned_data, variable_to_use,
  levels_to_use = NULL, k = 0)
}
\arguments{
\item{binned_data}{Should be a dataframe in binned format.}

\item{variable_to_use}{A string that indicates what variable to analyze.}

\item{levels_to_use}{A single string or a vector of strings.}

\item{k_value}{Returns indices for all sites (e.g. neurons) that have at least
k presentations of each condition.}
}
\value{
\item{inds_of_sites_with_at_least_k_repeats}{The indices of sites that have at least 
      k repetitions of each condition.}
\item{min_num_repeats_all_sites}{This parameter lists for all sites the number of 
     repetitions present for the label thathas the minium number of repetitions.}
}
\description{
This function takes in labels in binned label format, and an integer k, 
and returns the indices for all sites (e.g. neurons) that have at least
k presentations of each condition.
}
\examples{

}
