% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_matlab_raster_data.R
\name{convert_matlab_raster_data}
\alias{convert_matlab_raster_data}
\title{A function that converts raster data in MATLAB .mat format to R .rda format}
\usage{
convert_matlab_raster_data(matlab_raster_dir_name,
  r_raster_dir_name = NULL, start_ind = NULL, end_ind = NULL,
  files_contain = "")
}
\description{
A function that converts raster data in MATLAB .mat format to R .rda format
}
\details{
! link to raster format and related two functions
}
