% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binned_data_from_matlab_raster_data.R
\name{create_binned_data_from_matlab_raster_data}
\alias{create_binned_data_from_matlab_raster_data}
\title{A function that create binned data in .Rda format from raster data in .mat format}
\usage{
create_binned_data_from_matlab_raster_data(matlab_raster_dir_name,
  save_prefix_name, bin_width, sampling_interval, start_ind = NULL,
  end_ind = NULL, files_contain = "")
}
\description{
! link to raster format and binned format and related two functions
}
