% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster_data_from_matlab_raster_data.R
\name{create_raster_data_from_matlab_raster_data}
\alias{create_raster_data_from_matlab_raster_data}
\title{A function that converts raster data in .mat format to .Rda format}
\usage{
create_raster_data_from_matlab_raster_data(matlab_raster_dir_name,
  r_raster_dir_name = NULL, start_ind = NULL, end_ind = NULL,
  files_contain = "")
}
\description{
! link to raster format and related two functions
}
