% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_standard.R
\name{cv_standard}
\alias{cv_standard}
\title{The standard cross-validator (CV)  object}
\usage{
cv_standard(datasource, classifier, feature_preprocessors,
  num_resample_runs = 50, result_metrics = NULL,
  test_only_at_training_time = FALSE)
}
\description{
A cross-validator object takes a datasource, feature preprocessors and a classifier
and runs multiple cross-validation cycles by getting new training and test data splits, 
running the preprocessor to do preprocessing of the data, trains and tests the classifier, and 
the creates metric to evaluation the classification performance on the test set.
}
