% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_zscore.R
\name{fp_zscore}
\alias{fp_zscore}
\title{A feature preprocessor (FP) that zscore normalizes the data}
\usage{
fp_zscore()
}
\description{
This feature preprocessor object find the mean and standard deviation using the training data. 
The proprocessor then z-score transforms the training and test data using this mean and standard deviation
by subtracting the mean and dividing by the standard deviation.
}
