% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{get_data}
\alias{get_data}
\title{Register the generic function get_data() for datasource (DS) objects
Need to fill in more complete documentation here about the DS interface}
\usage{
get_data(ds_obj)
}
\arguments{
\item{ds_obj}{The data source object}
}
\description{
Register the generic function get_data() for datasource (DS) objects
Need to fill in more complete documentation here about the DS interface
}
