% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_label_repetitions.R
\name{get_num_label_repetitions_each_site}
\alias{get_num_label_repetitions_each_site}
\title{Get the number of trial repetitions for a given label for each site}
\usage{
get_num_label_repetitions_each_site(
  binned_data,
  label_to_assess,
  label_levels = NULL
)
}
\arguments{
\item{binned_data}{A string that list a path to a file that has data in
binned format, or a data frame of binned_data that is in binned format.}

\item{label_to_assess}{A string specifying which label variable should be
used for calculating the minimum number of level repetitions.}

\item{label_levels}{A character vector specifying which levels to include.
If not set, all levels will be used.}
}
\description{
Calculates how many repeated trials there are for each label
level for each site. This can be useful for selecting sites that have a
minimum number of repetitions of each stimulus or other experimental
condition.
}
\note{
The returned value is an S3 object that inherits from data.frame that
has an associated plot() method.
}
