% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{get_predictions}
\alias{get_predictions}
\title{Register the generic function get_predictions for the classifier (CL) objects}
\usage{
get_predictions(cl_obj, training_set, test_set)
}
\arguments{
\item{cl_obj}{The classifier object}

\item{training_set}{The training data}

\item{test_set}{The test data from all times}
}
\description{
Register the generic function get_predictions for the classifier (CL) objects
}
