% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_raster_data.R
\name{load_csv_raster_data}
\alias{load_csv_raster_data}
\title{Load a csv file into raster_data format}
\usage{
load_csv_raster_data(raster_file_name)
}
\arguments{
\item{raster_file_name}{A string specifying the name (and path) to a csv that
has the appropriate raster data column names (i.e., columns that start with
site.info, labels. and time.)}
}
\description{
Loads csv (comma separated value) files that has the appropriate raster_data
column names (i.e., columns that start with site.info, labels. and time.),
and returns data in raster_data format (i.e., a data frame with the raster.data
class attribute).
}
\examples{

\dontrun{

csv_raster_file_name <- file.path(
  system.file("extdata", package = "NeuroDecodeR"),
  "Zhang_Desimone_7object_raster_data_small_csv",
  "bp1001spk_01A_raster_data.csv"
)

# load the csv file into a raster_data data frame
raster_data <- load_csv_raster_data(csv_raster_file_name)

}

}
