% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_and_save_decoding_results.R
\name{log_save_results}
\alias{log_save_results}
\title{A function that saves the DECODING_RESULTS and logs the parameters used in the analysis}
\usage{
log_save_results(
  DECODING_RESULTS,
  save_directory_name,
  result_name = "No result name set"
)
}
\arguments{
\item{DECODING_RESULTS}{A list of results returned by the cross-validator's
run_decoding method.}

\item{save_directory_name}{A directory where the decoding results should be
saved.}

\item{result_name}{A string that gives a human readable name for the results 
that are to be saved. The default value is "No result name set".}
}
\description{
This function takes results returned by the cross-validator's run_decoding() 
 method and uses the cross-validator's get_properties() method to save a 
 log of the results that be used to reload specific results
}
