% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_correlation_CL.R
\name{max_correlation_CL}
\alias{max_correlation_CL}
\title{A maximum correlation coefficient classifier (CL) object}
\usage{
max_correlation_CL()
}
\description{
An implementation of a maximum correlation coefficeint classifier. 
Like all classifiers (CL) objects, this classifier has a get_predictions()
method which learns a model based on training data and then makes 
predictions on the test data.
}
