% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDTr.R
\docType{package}
\name{NDTr}
\alias{NDTr}
\alias{NDTr-package}
\title{NDTr: A package for doing neural decoding}
\description{
The NDTr is a package that makes it easy to do neural decoding analyses.
}
\details{
The NDTr is built around 5 abstract object types that work together in a
modular way to allow a range of neural decoding analyses. These five object
types are:
}
\section{1. Datasources}{

 Datasources take
}

\section{2. Feature preprofessors}{

 XXX
}

\section{3. Classifiers}{

 XXX
}

\section{4. Result metrics}{
 
 XXX
}

\section{5. Cross-validators}{

 XXX
}

\section{Data formats}{
 
 XXX
}

