% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_confusion_matrix.R
\name{plot.rm_confusion_matrix}
\alias{plot.rm_confusion_matrix}
\title{plot the confusion matrix results}
\usage{
\method{plot}{rm_confusion_matrix}(confusion_matrix_obj)
}
\arguments{
\item{confusion_matrix_obj}{A rm_confusion_matrix object that has aggregated
runs from a decoding analysis, e.g., if DECODING_RESULTS are the out from
the run_decoding(cv) then this argument should be
DECODING_RESULTS$rm_confusion_matrix.}
}
\description{
This function plots confusion matrices after the decoding analysis has been
run (and all results have been aggregated)
}
\seealso{
Other result_metrics: \code{\link{plot.rm_main_results}},
  \code{\link{plot_MI.rm_confusion_matrix}},
  \code{\link{rm_confusion_matrix}},
  \code{\link{rm_main_results}}
}
\concept{result_metrics}
