% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot_MI}
\alias{plot_MI}
\title{Plot the mutual information computed from a confusion matrix}
\usage{
plot_MI(rm_obj, plot_type)
}
\arguments{
\item{rm_obj}{the result metric object.}

\item{plot_type}{The type of plot that should be displayed.}
}
\description{
Plot the mutual information computed from a confusion matrix
}
