% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_confusion_matrix.R
\name{plot_MI.rm_confusion_matrix}
\alias{plot_MI.rm_confusion_matrix}
\title{plot the mutual information computed from a confusion matrix}
\usage{
\method{plot_MI}{rm_confusion_matrix}(rm_obj, plot_type = "TCD")
}
\arguments{
\item{rm_obj}{A rm_confusion_matrix object that has aggregated
runs from a decoding analysis, e.g., if DECODING_RESULTS are the out from
the run_decoding(cv) then this argument should be
DECODING_RESULTS$rm_confusion_matrix.}

\item{plot_type}{A string specifying the type of results to plot. Options are
'TCD' to plot a temporal cross decoding matrix or 'line' to create a line
plot of the decoding results as a function of time}
}
\description{
This function can plot line results or temporal cross-decoding results for
the the zero-one loss, normalized rank and/or decision values after the
decoding analysis has been run (and all results have been aggregated)
}
\seealso{
Other result_metrics: 
\code{\link{plot.rm_confusion_matrix}()},
\code{\link{plot.rm_main_results}()},
\code{\link{rm_confusion_matrix}()},
\code{\link{rm_main_results}()}
}
\concept{result_metrics}
