% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Register the generic function preprocess_data() for feature preprocessor (FP) objects
Need to fill in more complete documentation here about the FP interface}
\usage{
preprocess_data(fp_obj, train_data, all_times_test_data)
}
\arguments{
\item{fp_obj}{The FP object}

\item{train_data}{The training data}

\item{all_times_test_data}{The test data}
}
\description{
Register the generic function preprocess_data() for feature preprocessor (FP) objects
Need to fill in more complete documentation here about the FP interface
}
