% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{run_decoding}
\alias{run_decoding}
\title{Register the generic function run_decoding
Need to fill in more complete documentation here about the CV interface}
\usage{
run_decoding(cv_obj)
}
\arguments{
\item{cv_obj}{The CV object}
}
\description{
Register the generic function run_decoding
Need to fill in more complete documentation here about the CV interface
}
