% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_manage_decoding_results.R
\name{save_and_log_results}
\alias{save_and_log_results}
\title{A function saves DECODING_RESULTS and logs the parameters used in the analysis}
\usage{
save_and_log_results(DECODING_RESULTS, save_directory_name)
}
\arguments{
\item{DECODING_RESULTS}{A list of results returned by the cross-validator's
run_decoding method.}

\item{save_directory_name}{A directory where the decoding results should be
saved.}
}
\description{
This function takes results returned by the cross-validator's run_decoding() 
 method and uses the cross-validator's get_properties() method to save a 
 log of the results that be used to reload specific results
}
