% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_k_features_FP.R
\docType{data}
\name{select_k_features_FP}
\alias{select_k_features_FP}
\title{A feature preprocessor (FP) that reduces the data to the best k features}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
select_k_features_FP
}
\description{
This feature prerpocessor object find the k most selective features using an ANOVA on the training data. 
The proprocessor then eleminates all other features in both the training and test sets. This preprocessor
can also eliminate the best k features. 
This object uses \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{R6 package}
}
\keyword{datasets}
