% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_valid_NDR_objects.R
\name{test_valid_binned_format}
\alias{test_valid_binned_format}
\title{Tests if a data frame is in valid binned format}
\usage{
test_valid_binned_format(binned_data)
}
\arguments{
\item{binned_data}{A data frame or string specifying a file that will be
checked to see if it is in valid binned format.}
}
\value{
Returns NULL if object is in valid binned format. Otherwise it will
give an error message.
}
\description{
This function takes a data frame and tests that the data frame is in valid
binned format by checking that the data frame contains variables with the
appropriate names. If the data frame is not in correct binned format, an
error will be thrown that contains a message why the data is not in valid
binned format.
}
\keyword{internal}
