% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_DS.R
\docType{data}
\name{basic_DS}
\alias{basic_DS}
\title{A basic datasource object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
basic_DS
}
\description{
A datasource object takes data in binned format and returns training 
and testing splits of the data that can be passed to a classifier. 
This object uses \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{R6 package}
}
\section{basic_DS constructor}{


\describe{
\item{\code{basic_DS$new(binned.data, specific.binned.label.name, num.cv.splits, use.count.data, num.times.to.repeat.labels.per.cv.block )}}{
if successful, will return a new \code{basic_DS} object.
}}
}

\keyword{datasets}
