% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_num_level_repetitions.R
\name{calc_num_level_repetitions}
\alias{calc_num_level_repetitions}
\title{calc_num_level_repetitions}
\usage{
calc_num_level_repetitions(binned_data, variable_to_use,
  levels_to_use = NULL, k = 0)
}
\description{
This function takes in labels in binned label format, and an integer k,
and returns the indices for all sites (e.g. neurons) that have at least
k presentations of each condition.
}
