% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_max_correlation.R
\name{cl_max_correlation}
\alias{cl_max_correlation}
\title{A maximum correlation coefficient classifier (CL)}
\usage{
cl_max_correlation(ndr_container_or_object = NULL)
}
\arguments{
\item{ndr_container_or_object}{The purpose of this argument is to make the
constructor of the cl_maximum_corrleation classifier work with magrittr. If
this is set to the default value of NULL, then the constructor will return
a cl_max_correlation ndr classifier object. If this is set to an ndr
container, then a cl_max_correlation ndr classifier object will be added to
the container and the container will be returned. If this argument is set
to another ndr object, then both that ndr object as well as a new
cl_maximum_corrleation classifier will be added to a new container and the
container will be returned.}
}
\description{
An implementation of a maximum correlation coefficient classifier.
}
\details{
This CL object learns a mean population vector (template) for each
class from the training set (by averaging together the all training points
within each class). The classifier is tested by calculated Pearson’s
correlation coefficient between a test point and the templates learned from
the training set, and the class with the highest correlation value is
returned as the predicted label. The decision values returned by the
classifier are the correlation coefficients between all test points and all
templates.

Like all classifiers (CL) objects, this classifier has a get_predictions()
method which learns a model based on training data and then makes predictions
on the test data.
}
\examples{
# running a basic decoding analysis using the cl_max_correlation
data_file <- system.file(file.path("extdata", "ZD_150bins_50sampled.Rda"),
                         package = "NeuroDecodeR")
ds <- ds_basic(data_file, "stimulus_ID", 18)
fps <- list(fp_zscore())

cl <- cl_max_correlation()
cv <- cv_standard(ds, cl, fps)
\dontrun{
DECODING_RESULTS <- run_decoding(cv)
}

}
\seealso{
Other classifier: 
\code{\link{cl_poisson_naive_bayes}()},
\code{\link{cl_svm}()}
}
\concept{classifier}
