% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_max_correlation.R
\name{cl_max_correlation}
\alias{cl_max_correlation}
\title{A maximum correlation coefficient classifier (CL) object}
\usage{
cl_max_correlation()
}
\description{
An implementation of a maximum correlation coefficeint classifier. 
Like all classifiers (CL) objects, this classifier has a get_predictions()
method which learns a model based on training data and then makes 
predictions on the test data.
}
