% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_poisson_naive_bayes.R
\name{cl_poisson_naive_bayes}
\alias{cl_poisson_naive_bayes}
\title{A Poisson Naive Bayes classifier (CL) object}
\usage{
cl_poisson_naive_bayes()
}
\description{
An implementation of a Poisson Naive Bayes classifier. Like all classifiers,
this classifier learning a model based on training data and then makes
predictions on new test data.
}
\details{
Note: this classifier uses spike counts, so the binned data must be converted
to use this classifier, for exmaple, if you are using the basic_DS data
source, then use_count_data = TRUE should be set in the contructor. Also,
preprocessors that convert the data into values that are not integers should
not be used, for example, the fp_zscore should not be used with this
classifier.
}
\examples{
 data_file <- system.file("extdata/ZD_150bins_50sampled.Rda", package = "NDTr") 
 ds <- ds_basic(data_file, 'stimulus_ID', 18, use_count_data = TRUE)
 fps <- list()
 cl <- cl_poisson_naive_bayes()
 cv <- cv_standard(ds, cl, fps) 
 

}
\seealso{
Other classifier: 
\code{\link{cl_max_correlation}()},
\code{\link{cl_svm}()}
}
\concept{classifier}
