% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binned_data_from_matlab_raster_data.R
\name{create_binned_data_from_matlab_raster_data}
\alias{create_binned_data_from_matlab_raster_data}
\title{A function that create binned data in .Rda format from raster data in .mat format}
\usage{
\code{create_binned_data(raster_dir_name, save_prefix_name, bin_width, sampling_interval, start_ind, end_ind, files_contain)}
}
\arguments{
\item{save_prefix_name}{character. Prefix to the generated name for the created binned file, which is
"\code{bin_width}_samples_binned_every_\code{sampling_interval}_samples}

\item{bin_width}{integer. The bin width over which raster data is averaged.}

\item{sampling_interval}{integer. It specifies the nth sample following the start of a bin, where the next bin starts}

\item{start_ind}{integer. It specifies the sample index where the binning process starts. Due to the structure of raster data in matlab, all sample indices should be positive. 
By default, all data are included.}

\item{end_ind}{integer. It specifies the sample index where the binning process should end by. Due to the structure of raster data in matlab, all sample indices should be positive.
By default, all data are included.}

\item{files_contain}{regular expression. Only raster data files that match the file_contains are inlcluded. By default, it is an empty character.}

\item{matlab_dir_name}{character. Name of a directory containing raster data in .mat format.}
}
\value{
Preceding the binning of each raster file, it spills the total number of raster files will have been binned as you will see
the number increments by one. After the creation of all files, it spills the binned file name. By default, it is an empty character.
}
\description{
! link to raster format and binned format and related two functions
}
\examples{
Bin the data using 150 sample bins sample at 50 sample intervals
Assumes that raster files are in the directory 'data/Zhang_Desimone_7objects_raster_data_mat/'
and saves the output file with the prefix ZD
\dontrun{
create_binned_data_from_matlab_raster_data(file.path(getwd(),'data/raster/Zhang_Desimone_7objects_raster_data_mat/'), 'data/binned/ZD', 150, 20)
}
If you get other files mixed in the raster directory that are .mat files and only want to include data from 200th sample to 800th sample
\dontrun{
create_binned_data_from_matlab_raster_data(file.path(getwd(),'data/raster/Zhang_Desimone_7objects_raster_data_mat/'), 'data/binned/ZD', 150, 20, 200, 800, "\\\\.mat$")
}
}
