% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster_data_from_matlab_raster_data.R
\name{create_raster_data_from_matlab_raster_data}
\alias{create_raster_data_from_matlab_raster_data}
\title{export}
\usage{
create_raster_data_from_matlab_raster_data(matlab_raster_directory_name)
}
\description{
export
}
