% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp_select_k_features.R
\name{fp_select_k_features}
\alias{fp_select_k_features}
\title{A feature preprocessor (FP) that reduces the data to the best k features}
\usage{
fp_select_k_features(num_site_to_use = NA, num_sites_to_exclude = NA)
}
\arguments{
\item{num_site_to_use}{The number of features with the smallest p-values to use.}

\item{num_sites_to_exclude}{The number of features with the smallest p-values
that should be excluded.}
}
\description{
This feature preprocessor object applies an ANOVA to the training data to
find the p-value of all features. It then either uses the top k features with
the smallest p-values, or it removes the features with the smallest k
p-values. Additionally, this function can be used to remove the top k
p-values and then use only the following j next smallest p-values (for
example, this can be useful if one is interesting in comparing the
performance using the most selective 10 neurons to using the next 10 most
selective neurons, etc.).
}
\examples{
# This will cause the cross-validator use only the 50 most selective sites
fp <- fp_select_k_features(num_site_to_use = 50)

# This will cause the cross-validator to remove the 20 most selective sites
fp <- fp_select_k_features(num_sites_to_exclude = 20)

# This will cause the cross-validator to remove the 20 most selective sites 
# and then use only the 50 most selective sites that remain after the 20 are 
# eliminated 
fp <- fp_select_k_features(50, 20)


}
\seealso{
Other feature_preprocessor: \code{\link{fp_zscore}}
}
\concept{feature_preprocessor}
