% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{get_parameters}
\alias{get_parameters}
\title{Return all parameters that an object uses to enable reproducible analyses}
\usage{
get_parameters(ndtr_obj)
}
\arguments{
\item{ndtr_obj}{The object from the NDTr package that one should get the parameters from}
}
\description{
Return all parameters that an object uses to enable reproducible analyses
}
