% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{get_predictions}
\alias{get_predictions}
\title{Register the generic function get_predictions for the classifier (CL) objects}
\usage{
get_predictions(obj, train_data, all_times_test_data)
}
\arguments{
\item{obj}{The CL object}

\item{train_data}{The training data}

\item{all_times_test_data}{The test data}
}
\description{
Register the generic function get_predictions for the classifier (CL) objects
}
