% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_correlation_CL.R
\docType{data}
\name{max_correlation_CL}
\alias{max_correlation_CL}
\title{A maximum correlation coefficient classifier (CL) object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
max_correlation_CL
}
\description{
An implementation of a maximum correlation coefficeint classifier. Like all classifiers, this classifier
learning a model based on training data and then makes predictions on new test data.  
This object uses \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{R6 package}
}
\section{max_correlation_CL constructor}{


\describe{
\item{\code{max_correlation_CL$new()}}{
if successful, will return a new \code{max_correlation_CL} object.
}}
}

\keyword{datasets}
