% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDTr.R
\docType{package}
\name{NDTr}
\alias{NDTr}
\title{NDTr: A package for neural decoding analyses}
\description{
The NDTr is a package that makes it easy to do neural decoding analyses.
}
\details{
The NDTr is built around 5 abstract object types that work together in a
modular way to allow a range of neural decoding analyses. These five object
types are:
\enumerate{
\item Datasources (DS): Generate training and test splits of the data.
\item Feature preprofessors (FP): Learn parameters on the training set and apply
transformations to the training and test sets.
\item Classifiers (CL): Learn the relationship between experimental conditions
(i.e., "labels") and neural data on a training set, and then predict
experimental conditions neural data in a test set.
\item Result metrics (RM): Aggregate results across validation splits and
over resampled runs and compute and plot final decoding accuracy metrics.
\item Cross-validators (CV): Take the DS, FP, CL and RM objects and run a
cross-validation decoding procedure.
}
}
\section{Data formats}{
Two data formats are used to do decoding analyses which are:
\enumerate{
\item \verb{raster format} contains high temporal precision data where neural
activity from each site is stored in a spearate file.
\item \verb{binned format} contains data from multiple sites and the data is more
coarsely binned across time.
}

A user of the NDTr will typically store their data in \verb{raster format} and
then use the \code{\link[=create_binned_data]{create_binned_data()}} to create a \verb{binned format} data file
that will be used in the decoding analysis.
}

