% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_label_repetitions.R
\name{plot.label_repetition}
\alias{plot.label_repetition}
\title{A plot function for label_repetition object}
\usage{
\method{plot}{label_repetition}(x, ..., show_legend = TRUE)
}
\arguments{
\item{x}{A label_repetition object that was generated from calling the
get_num_label_repetitions() function.}

\item{...}{This is needed to conform to the plot generic interface.}

\item{show_legend}{A Boolean specifying whether to show a legend
that list which label each color in the plot corresponds to.}
}
\value{
. Returns a ggplot object that plots the number of sites that
that have at least k label repetitions as a function of k.
}
\description{
This function plots the number of sites have at least k label repetitions.
Creating this plot is useful for assessing how to set the number of
cross-validation splits (and repeats of labels per cross-validation split) to
use in a datasource. This function returns a ggplot2 object which can be
further modified as needed.
}
