% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{plot_MI}
\alias{plot_MI}
\title{Plot the mutual information computed from a confusion matrix}
\usage{
plot_MI(obj, plot_type)
}
\arguments{
\item{obj}{the RM object}

\item{plot_type}{The type of plot that should be displayed}
}
\description{
Plot the mutual information computed from a confusion matrix
}
