% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_generic_functions.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Register the generic function preprocess_data() for feature preprocessor (FP) objects
Need to fill in more complete documentation here about the FP interface}
\usage{
preprocess_data(fp, training_set, test_set)
}
\arguments{
\item{fp}{The FP object}

\item{training_set}{The training set for a particular time}

\item{test_set}{The test set from all times}
}
\description{
Register the generic function preprocess_data() for feature preprocessor (FP) objects
Need to fill in more complete documentation here about the FP interface
}
