% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_main_results.R
\name{rm_main_results}
\alias{rm_main_results}
\title{A result metric (RM) that calculates central decoding accuracy measures}
\usage{
rm_main_results()
}
\description{
This result metric calculate the zero-one loss, the normalized rank, and the
 mean of the decision values. This is also an S3 object which has an associated
 plot function to display the results.
}
\details{
Like all result metrics, this result metric has functions to aggregregate
results after completing each set of cross-validation classifications, and
also after completing all the resample runs. The results should then be
available in the DECODING_RESULTS object returned by the cross-validator.
}
\examples{
# This result metric does not take any arguments.
# If you only want to use the rm_main_results(), then you can put it in a
# list by itself and pass it to the cross-validator.
the_rms <- list(rm_main_results())

}
\seealso{
Other result_metrics: \code{\link{plot.rm_confusion_matrix}},
  \code{\link{plot.rm_main_results}},
  \code{\link{plot_MI.rm_confusion_matrix}},
  \code{\link{rm_confusion_matrix}}
}
\concept{result_metrics}
