% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_CV.R
\docType{data}
\name{standard_CV}
\alias{standard_CV}
\title{The standard cross-validator (CV)  object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
standard_CV
}
\description{
A cross-validator object takes a datasource, feature preprocessors and a classifier
and runs multiple cross-validation cycles by getting new training and test data splits, 
running the preprocessor to do preprocessing of the data, trains and tests the classifier, and 
the creates metric to evaluation the classification performance on the test set.  
This object uses \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{R6 package}
}
\section{standard_CV}{


\describe{
\item{\code{standard_CV$new(data_source, classifier, feature_preprocessors)  )}}{
If successful, will return a new \code{basic_DS} object.
}}
}

\keyword{datasets}
