% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_FP.R
\docType{data}
\name{zscore_FP}
\alias{zscore_FP}
\title{A feature preprocessor (FP) that zscore normalizes the data}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
zscore_FP
}
\description{
This feature prerpocessor object find the mean and standard deviation using the training data. 
The proprocessor then z-score transforms the training and test data using this mean and standard deviation
by subtracting the mean and dividing by the standard deviation. 
This object uses \href{https://cran.r-project.org/web/packages/R6/vignettes/Introduction.html}{R6 package}
}
\section{zscore_FP constructor}{


\describe{
\item{\code{zscore_FP$new()}}{
if successful, will return a new \code{zscore_FP} object.
}}
}

\keyword{datasets}
